Remove send button for now.
The implementation doesn't work well on iOS (causes keyboard to hide /
flicker away), and in general reduces the space for entry text. It might
be possible to fix the keyboard hide/show issues by placing buttons
below the entry and checking the position of touches, but I don't think
it's worth the effort for now.