Fix notification upgrade heisenbug.
In tracking this down, I was looking for a way that a message could be
marked in between a message's "new-message" mark and its "new-mention"
mark. It turns out that in the initial message load, when all message
changes are handled in a batch, the marks are added in the order of
their filters, so "new-reply" shows up first.

The alerts mapping no longer had a "new-message" alert because a
subsequent message (which was a reply) had evicted it, and then the
"new-message" was replaced with a "new-reply". The alert replacement
code assumed that even if an alert was replaced, it would still exist,
which is not the case when a "new-message" turns into a "new-reply".

Instead of overcomplicating the loop with deduplicating alerts by
priority, we can just perform the only alert deduplication that has any
effect by handling mention alerts specifically.