Filter out no-longer-available messages from toolbox (for now).
It would be good to keep these around in the UI somehow, but currently
doing so is a bit tricky, and the resulting UX is difficult to explain.
I think in the future we'll keep more history in this area, which could
solve the problem in another way. Alternatively, could manually fetch
information specifically about the messages that are selected, once
there is an API to do so.