Fix backspace key in React 0.14.
It looks like the order in which event handlers fire changed in the
React upgrade. This is a bandaid that fixes an oversight in the original
code (the synthetic event target not being set). The input handling code
is in need of some restructuring that's outside the scope of this fix
for the moment.