Buffer outgoing WebSocket packets when not connected
This is a bandaid for exceptions in which the UI wants to send but the
socket isn't ready at the moment. Currently, loadMoreLogs is a major
culprit. These issues should probably be addressed by fixing cases where
the UI state needs to wait until a connection is available, but it's
much easier to paper over the issue here for now.