Band-aid for account settings store handling login popup failure
These API calls should turn into promises so that their flow is tracked
specifically rather than global events. In this case, a flow event was
being handled by a store when it shouldn't have been, but the store
itself doesn't have enough state to definitively tell when it should be
listening.