Fix user list bug
If there is only one user, they would not be displayed when using the
development frontend.
It has very probably to do with the deduplication, and might be related
to repeated evaluation of the (stateful!) filter, so that the single
entry becomes (erroneously) deduplicated upon repeated iterations; an
eager (and therefore single) evaluation of the filter is forced using
the toList().
The behavior is for unknown reasons not observed in the production
frontend.