Fix page location reading in Chrome
Chrome started setting the location of the iframe containing the bulk
of the client code slightly differently to that one of the top-level
window in some unknown version; the iframe does now not contain the
fragment identifier of the top-level window, which some things
(including custom themes and development the frontend against upstream
backends) depended on.

Ironically, a follow-up to 3ba8a2ec50efedbce41723392b7f9abe7ac5976f.